Update time tracker test
